% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{fix.csv}
\alias{fix.csv}
\title{Use the data editor for a CSV file}
\usage{
fix.csv(file, new.name = TRUE)
}
\arguments{
\item{file}{string of filename}

\item{new.name}{T/F}
}
\description{
This function loads a CSV file, lets the user edit it in the native data
editor, then re-saves it, prompting the user for a new file name if desired.
}
