% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1000x.R
\name{load_1000x}
\alias{load_1000x}
\title{Title}
\usage{
load_1000x(
  path,
  datetime = "TIMESTAMP",
  timezone = "EST",
  col_names = NA,
  skip_4_names = 1,
  skip_4_dat = 3
)
}
\arguments{
\item{path}{A path to the data readout to be loaded.}

\item{datetime}{String to be used for the datetime column.}

\item{timezone}{POSIXct timezone for readout.}

\item{col_names}{Vector of column names same length as number of columns in the readout.}

\item{skip_4_names}{Number of lines to skip in the readout to get to column names.}

\item{skip_4_dat}{Number of lines to skip in the readout to get data.}
}
\value{

}
\description{
Title
}
\examples{
ex_path <- system.file('extdata', 'treefort_1000x.dat', package = 'wxlogR')
df <- load_1000x(ex_path)
}
