% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_CS.R
\name{load_CS_10}
\alias{load_CS_10}
\title{Load Campbell Sci 10x logger readout.}
\usage{
load_CS_10(
  path,
  timezone = "EST",
  col_names = NA,
  pos_year = 2,
  pos_doy = 3,
  pos_time = 4,
  index = 115
)
}
\arguments{
\item{path}{A path to the data readout to be loaded.}

\item{timezone}{POSIXct timezone for readout.}

\item{col_names}{Vector of column names same length as number of columns in the readout.}

\item{pos_year}{Column position for year.}

\item{pos_doy}{Column position for day of year.}

\item{pos_time}{Column position for CS time. Formatted as (ex. 100 == 01:00:00 a.m. 60 minutes since midnight)}

\item{index}{what is the row index you are after. Default is 115 which is the 15 min average and removes the daily average.}
}
\value{
logger readout with datetime parsed and with column headers and daily average removed
}
\description{
Load Campbell Sci 10x logger readout.
}
\examples{
ex_path <- system.file('extdata', '10x.dat', package = 'wxlogR')
df <- load_CS_10(ex_path)
}
