% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_CS_time.R
\name{parse_CS_time}
\alias{parse_CS_time}
\title{Parse Campbell Sci 10x time.}
\usage{
parse_CS_time(time)
}
\arguments{
\item{time}{formatted as (ex. 100 == 01:00:00 a.m. 60 minutes since midnight)}
}
\value{
time integer formatted so that posixct can accept (HHMMSS)
}
\description{
Parse Campbell Sci 10x time.
}
\examples{
parse_CS_time('100')
}
