% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{plot_data_gaps}
\alias{plot_data_gaps}
\title{Return Plot Showing Presence of Datetime Value as 1}
\usage{
plot_data_gaps(datetime, from = NA, to = NA)
}
\arguments{
\item{datetime}{POSIXct datetime vector with potential gaps}

\item{from}{POSIXct or Date or NA for min(datetime)}

\item{to}{POSIXct or Date or NA for max(datetime}
}
\value{
ggplot
}
\description{
Return Plot Showing Presence of Datetime Value as 1
}
\examples{
ex_path <- system.file('extdata', 'treefort_1000x.dat', package = 'wxlogR')
df <- load_CS_1000(ex_path)
plot_data_gaps(df$TIMESTAMP)
}
