% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qc_data_filter}
\alias{qc_data_filter}
\title{Sets new values for where flags exist}
\usage{
qc_data_filter(df, col_val, col_flag, good_flags, flag_val)
}
\arguments{
\item{df}{dataframe}

\item{col_val}{string or vector of strings referring to dataframe column of the raw values to be updated.}

\item{col_flag}{string referring to the dataframe column of flags.}

\item{good_flags}{vector of acceptable flags, type is same as the col_flag col.}

\item{flag_val}{value to insert in col_val where is flagged.}
}
\value{
NA, updates existing col_val column in df.
}
\description{
Sets new values for where flags exist
}
