% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_CS.R
\name{load_CS_1000}
\alias{load_CS_1000}
\title{Load Campbell Sci formatted logger readout. Default is setup to work with 1000x / 3000x format.}
\usage{
load_CS_1000(
  path,
  datetime = "TIMESTAMP",
  timezone = "EST",
  col_names = NA,
  skip_4_names = 1,
  skip_4_dat = 3
)
}
\arguments{
\item{path}{A path to the data readout to be loaded.}

\item{datetime}{String to be used for the datetime column.}

\item{timezone}{POSIXct timezone for readout.}

\item{col_names}{Vector of column names same length as number of columns in the readout.}

\item{skip_4_names}{Number of lines to skip in the readout to get to column names.}

\item{skip_4_dat}{Number of lines to skip in the readout to get data.}
}
\value{

}
\description{
Load Campbell Sci formatted logger readout. Default is setup to work with 1000x / 3000x format.
}
\examples{
ex_path <- system.file('extdata', 'treefort_1000x.dat', package = 'wxlogR')
df <- load_CS_1000(ex_path)
}
