% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy_cov.R
\name{load_CS_HF}
\alias{load_CS_HF}
\title{Load Campbell Sci hi frequency (HF) data, if header is TOB3 need to convert
to TOA5 using card convert. Uses data.table package for loading for increased
performance. Should load 5 gigs in a few seconds.}
\usage{
load_CS_HF(
  path,
  datetime = "TIMESTAMP",
  timezone = "EST",
  col_names = NA,
  skip_4_names = 1,
  skip_4_dat = 4,
  num_rows = 1000
)
}
\arguments{
\item{path}{A path to the data readout to be loaded.}

\item{datetime}{String to be used for the datetime column.}

\item{timezone}{The POSIXct timezone for original readout.}

\item{col_names}{Vector of column names same length as number of columns in the readout.}

\item{skip_4_names}{Number of lines to skip in the readout to get to column names. Default = 1.}

\item{skip_4_dat}{Number of lines to skip in the readout to get data. Default = 3.}

\item{num_rows}{Number of rows to retrieve.}
}
\value{

}
\description{
Load Campbell Sci hi frequency (HF) data, if header is TOB3 need to convert
to TOA5 using card convert. Uses data.table package for loading for increased
performance. Should load 5 gigs in a few seconds.
}
\examples{
ex_path <- system.file('extdata', 'HF_EC_example.dat', package = 'wxlogR')
df <- load_CS_HF(ex_path)
}
